requirements.txt
Flask>=3.0,<4.0
langgraph
langchain
langchain-core
streamlit
chromadb
python-dotenv
requests
